+++ internal/server/adapters/storage/database/postgres.go
@@ -5,9 +5,13 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"metrics/internal/server/config"
 	"metrics/internal/server/core/domain"
 	"metrics/internal/server/logger"
 
+	"github.com/avast/retry-go"
+	"github.com/jackc/pgerrcode"
+	"github.com/jackc/pgx/v5/pgconn"
 	_ "github.com/jackc/pgx/v5/stdlib"
 	"github.com/jmoiron/sqlx"
 	"go.uber.org/zap"
@@ -194,4 +198,33 @@ func (s *MetricStorage) Ping(ctx context.Context) error {
 		return fmt.Errorf("failed to ping database %w", err)
 	}
 	return nil
+}
+
+func (s *MetricStorage) retryExecRequest(ctx context.Context, query string, args ...any) error {
+	var originalErr error
+	err := retry.Do(
+		func() error {
+			_, originalErr := s.db.ExecContext(
+				ctx,
+				query,
+				args...,
+			)
+			return originalErr
+		},
+		retry.RetryIf(func(err error) bool {
+			var pgErr *pgconn.PgError
+			if errors.As(err, &pgErr) && pgerrcode.IsConnectionException(pgErr.Code) {
+				return true
+			}
+			return false
+		}),
+		retry.Attempts(config.Attempts),
+		retry.DelayType(config.DelayType),
+		retry.OnRetry(config.OnRetry),
+	)
+	if err != nil {
+		logger.Log.Error("retryError", zap.Error(err), zap.Error(originalErr))
+		return fmt.Errorf("%w", originalErr)
+	}
+	return originalErr
 }
